+++ issuer-service/src/test/java/ch/admin/bj/swiyu/issuer/service/OAuthServiceTest.java
@@ -1,12 +1,15 @@
 package ch.admin.bj.swiyu.issuer.service;
 
 import ch.admin.bj.swiyu.issuer.common.config.ApplicationProperties;
+import ch.admin.bj.swiyu.issuer.common.exception.OAuthError;
 import ch.admin.bj.swiyu.issuer.common.exception.OAuthException;
 import ch.admin.bj.swiyu.issuer.domain.credentialoffer.*;
 import ch.admin.bj.swiyu.issuer.service.webhook.EventProducerService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.Mockito;
 
 import java.util.Optional;
@@ -14,6 +17,7 @@ import java.util.Set;
 import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.Assert.assertThrows;
 
 class OAuthServiceTest {
 
@@ -72,8 +76,6 @@ class OAuthServiceTest {
                 .isNotEqualTo(refreshToken.toString());
         // verify that repository save was called (tokens updated)
         Mockito.verify(credentialManagementRepository).save(mockMgmt);
-
-
     }
 
     @Test
@@ -97,5 +99,31 @@ class OAuthServiceTest {
                 .isEqualTo(refreshToken.toString());
         // verify that repository save was called (tokens updated)
         Mockito.verify(credentialManagementRepository).save(mockMgmt);
+    }
+
+    @Test
+    void refreshOAuthToken_whenInvalidToken_thenThrowError() {
+        var exception = assertThrows(OAuthException.class, () -> oauthService.refreshOAuthToken("invalid token"));
+        assertThat(exception.getError()).isEqualTo(OAuthError.INVALID_REQUEST);
+    }
+
+    @Test
+    void getUnrevokedCredentialOfferByRefreshToken_whenInvalidRefreshToken_thenThrowError() {
+        var exception = assertThrows(OAuthException.class, () -> oauthService.getUnrevokedCredentialOfferByRefreshToken("invalid token"));
+        assertThat(exception.getError()).isEqualTo(OAuthError.INVALID_REQUEST);
+    }
+
+    @ParameterizedTest
+    @ValueSource(strings = {"bearer 0c16dd9c-1dcd-4fc1-b503-bc42c505f113", "BEARER 0c16dd9c-1dcd-4fc1-b503-bc42c505f113", "dpop 0c16dd9c-1dcd-4fc1-b503-bc42c505f113", "DPoP 0c16dd9c-1dcd-4fc1-b503-bc42c505f113"})
+    void getAccessToken_whenCorrectAuthorizationHeader_thenSuccess(String authorizationRequestHeader) {
+        final String ACCESS_TOKEN = "0c16dd9c-1dcd-4fc1-b503-bc42c505f113";
+        var extractedToken = oauthService.getAccessToken(authorizationRequestHeader);
+        assertThat(extractedToken).as("extracted token should match the access token").isEqualTo(ACCESS_TOKEN);
+    }
+
+    @ParameterizedTest
+    @ValueSource(strings = {"", "bearer ", "dpop ", "barer 0c16dd9c-1dcd-4fc1-b503-bc42c505f113", "0c16dd9c-1dcd-4fc1-b503-bc42c505f113"})
+    void getAccessToken_whenIllegalAuthorizationHeader_thenOAuthException(String authorizationRequestHeader){
+        assertThrows(OAuthException.class, () -> oauthService.getAccessToken(authorizationRequestHeader));
     }
 }